Defines class is compatible with QML

+++ sprout-app/src/main.cpp
@@ -5,7 +5,7 @@
 int main(int argc, char* argv[]) {
     if (argc != 2) {
         auto version = Defines::version();
-        qDebug() << qPrintable(QString("Sprout %1.%2.%3").arg(version["major"]).arg(version["minor"]).arg(version["patch"]));
+        qDebug() << qPrintable(QString("Sprout %1.%2.%3").arg(version["major"].toInt()).arg(version["minor"].toInt()).arg(version["patch"].toInt()));
         qDebug() << "Usage: sprout [sourcefile]";
     } else {
         QString filePath = argv[1];
